readings/get_random_ddrs/requirements.txt
python-dotenv
pandas
azure-cosmos